[#2697] Move prettifyCoordinates and prettifyProjection to Map directory.
This issue was identified in code review.
