+++ lib/Map/prettifyCoordinates.js
