+++ lib/Map/prettifyProjection.js
