+++ lib/Models/LocationMarkerUtils.js
@@ -6,7 +6,7 @@ import Cartographic from 'terriajs-cesium/Source/Core/Cartographic';
 import Cartesian3 from 'terriajs-cesium/Source/Core/Cartesian3';
 import VerticalOrigin from 'terriajs-cesium/Source/Scene/VerticalOrigin';
 import sampleTerrain from 'terriajs-cesium/Source/Core/sampleTerrain';
-import prettifyCoordinates from './prettifyCoordinates';
+import prettifyCoordinates from '../Map/prettifyCoordinates';
 
 import markerIcon from '../../wwwroot/images/map-pin.png';
